[lib/prep.sh] Limit calls to `cargo install --list` for plugins.
When checking for an installing plugins, there isn't a need to call
cargo for each plugin to check. While this is just-as-good as previous
behavior, it also limits any issues when calling a Cargo subcommand that
might fail.

Presently, there are times where a race exists and either `STDOUT` or
`STDERR` might disppear when Rustup (or potentially Cargo) is used in a
shell pipeline, leading to a program panic with `failed printing to
stdout: Broken pipe (os error 32)`. This change tries to minimize the
probability that this occurs until a more robust upstream fix has
landed.

References: rust-lang/rustup.rs#1730
References: rust-lang/rustup.rs#1765

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>