[prep] Add a temp file cleanup system via trap handlers.
The trap implementation is heavily inspired by a StackExchange reply
which produces a very shell-portable implementation that tries to
minimize the number of times that the trap action function is invoked.

For more details on the trap implementation, see:
https://unix.stackexchange.com/a/240736

The `cleanup_file` function will track a file for each invocation of the
function in a temp file which will be used in the trap handler
`trap_cleanup`. As this provides a convenient way to clean up work or
temporary files however the program terminates, the end result feels a
bit like a `defer` in Go, especially when called directly after creating
a temp file:

```
my_work() {
  tempwork="$(mktemp)"
  cleanup_file "$tempwork"
}
```

Note that there is a very trivial implementation of `mktemp_file`, but
this is only the result of a lot of reading and testing to find a
maximally portable way to create a temp file in a temp directory. If
this solution ends up not working on a platform, then it can be
altered without causing a refactoring of all the call sites. In
short--this should make the code easier to manage in the long run across
more platforms.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>