Re-implement `indent` to preserve exit code of command.
In minimal/POSIX shells there is no support for `set -o pipefail` which
means that the exit code of the first command in a shell pipeline won't
be addressable in an easy way. This implementation uses a temp file to
ferry the original command's exit code from a subshell back into the
main function. The output can be aligned with a pipe to `sed` as before
but now we have an implementation which mimicks a `set -o pipefail`
which should work on all Bourne shells. Note that the `set -o errexit`
is disabled during the command's invocation so that its exit code can be
captured.

Based on implementation from: https://stackoverflow.com/a/54931544

Rather than relying on `STDIN` coming in from a pipe, this version calls
the command to be indented and manages things from there. Therefore,
each `indent` call site is updated from the former pipe form:

```
echo hello | indent
```

to the prefix form:

```
indent echo hello
```

As the `indent` function internally redirects `STDERR` to `STDOUT` to
indent both streams, several old redirections of `2>&1` could be removed
at the `indent` call sites.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>