[darwin] Implement package install check caching.
This change is aimed to speed up the execution time when checking for
packages. Prior to this change, each call to install a package on macOS
would call some form of Homebrew's `brew list`/`brew cask list` in order
to grep the target package. In practice, Homebrew's CLI execution time
was starting to dominate the run time of `prep` itself. By saving off
the result of `brew list --version` & `brew cask list --versions` a
sample run of `prep` on a mostly configured macOS system went from 127
seconds to 72 seconds with only this optimization.

For completeness, the App Store installation logic was also augmented to
provide one full solution across macOS for a basis of comparison with
other platforms.

Note that this implementation uses a shared cache file which is
explicitly passed to each invocation. It's possible that future work
might eliminate this external setup and internalize the handling of this
cache file. At present however, this implementation has the benefit of
being explicit and fairly straight forward.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>