[ruby] Update stable Ruby versions automatically
This change updates the strategy for using ruby-install to install Ruby
versions. Prior to this change the stable version of Ruby wouldn't be
updated and built but rather the build would be skipped if at least one
Ruby version was found in `~/.rubies`.

Now `ruby-install --latest` is called each time to update the internal
versions of Rubies and `ruby-install --no-reinstall ruby` is called
which will skip a build if the current stable version of Ruby is
detected. Coolbeans.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>