[node] Update LTS Node versions automatically
This change updates the strategy for using nvm to install Node versions.
Prior to this change the LTS version of Node wouldn't be updated but
rather any installation would be skipped if at least one Node version
was found in `~/.nvm/versions/node`.

Now `nvm install --lts` is called each time which will skip an
installation if the current LTS version of Node is detected.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>